MeasureTheory

Introduction

API

MeasureTheory.Density

MeasureTheory.DensityMeasure

MeasureTheory.Kernel

MeasureTheory.PowerMeasure

MeasureTheory.SuperpositionMeasure

MeasureTheory.WeightedMeasure

MeasureTheory.:≃

MeasureTheory.:≪

MeasureTheory.basemeasure

MeasureTheory.isprimitive

MeasureTheory.logdensity

MeasureTheory.@measure

MeasureTheory.Density — Type

struct Density{M,B}
    μ::M
    base::B
end

For measures μ and ν with μ≪ν, the density of μ with respect to ν (also called the Radon-Nikodym derivative dμ/dν) is a function f defined on the support of ν with the property that for any measurable a ⊂ supp(ν), μ(a) = ∫ₐ f dν.

Because this function is often difficult to express in closed form, there are many different ways of computing it. We therefore provide a formal representation to allow comptuational flexibilty.

MeasureTheory.DensityMeasure — Type

struct DensityMeasure{F,B} <: AbstractMeasure
    density :: F
    base    :: B
end

A DensityMeasure is a measure defined by a density with respect to some other "base" measure

MeasureTheory.Kernel — Type

kernel(f, M)
kernel((f1, f2, ...), M)

A kernel κ = kernel(f, m) returns a wrapper around a function f giving the parameters for a measure of type M, such that κ(x) = M(f(x)...) respective κ(x) = M(f1(x), f2(x), ...)

If the argument is a named tuple (;a=f1, b=f1), κ(x) is defined as M(;a=f(x),b=g(x)).

Reference

https://en.wikipedia.org/wiki/Markov_kernel

MeasureTheory.PowerMeasure — Type

struct PowerMeasure{M,N}
    μ::M
    size::NTuple{N,Int}
end

A power measure is a product of a measure with itself. The number of elements in the product determines the dimensionality of the resulting support.

Note that power measures are only well-defined for integer powers.

The nth power of a measure μ can be written μ^x.

MeasureTheory.SuperpositionMeasure — Type

struct SuperpositionMeasure{X,NT} <: AbstractMeasure
    components :: NT
end

Superposition of measures is analogous to mixture distributions, but (because measures need not be normalized) requires no scaling.

The superposition of two measures μ and ν can be more concisely written as μ + ν.

MeasureTheory.WeightedMeasure — Type

struct WeightedMeasure{R,M} <: AbstractMeasure
    logweight :: R
    base :: M
end

MeasureTheory.:≃ — Method

≃(μ,ν)

Equivalence of Measure

Measures μ and ν on the same space X are equivalent, written μ ≃ ν, if μ ≪ ν and ν ≪ μ.

Note that this is often written ~ in the literature, but this is overloaded in probabilistic programming, so we use alternate notation.

MeasureTheory.:≪ — Function

≪(μ,ν)

Absolute continuity

The following are equivalent:

μ ≪ ν

μ is absolutely continuous wrt ν

There exists a function f such that μ = ∫f dν

MeasureTheory.basemeasure — Function

basemeasure(μ)

MeasureTheory.isprimitive — Method

isprimitive(μ)

Most measures are defined in terms of other measures, for example using a density or a pushforward. Those that are not are considered (in this library, it's not a general measure theory thing) to be primitive. The canonical example of a primitive measure is Lebesgue(X) for some X.

The default method is isprimitive(μ) = false

So when adding a new primitive measure, it's necessary to add a method for its type that returns true.

MeasureTheory.logdensity — Function

logdensity(μ::Measure{X}, x::X)

Compute the logdensity of the measure μ at the point x. This is the standard way to define logdensity for a new measure. the base measure is implicit here, and is understood to be basemeasure(μ).

Methods for computing density relative to other measures will be

MeasureTheory.@measure — Macro

@measure <declaration>

The <declaration> gives a measure and its default parameters, and specifies its relation to its base measure. For example,

@measure Normal(μ,σ) ≃ Lebesgue{X}

declares the Normal is a measure with default parameters μ and σ, and it is equivalent to its base measure, which is Lebesgue{X}

You can see the generated code like this:

julia> MacroTools.prettify(@macroexpand @measure Normal(μ,σ) ≃ Lebesgue{X})
quote
    struct Normal{P, X} <: AbstractMeasure
        par::P
    end
    function Normal(nt::NamedTuple)
        P = typeof(nt)
        return Normal{P, eltype(Normal{P})}
    end
    Normal(; kwargs...) = Normal((; kwargs...))
    (basemeasure(μ::Normal{P, X}) where {P, X}) = Lebesgue{X}
    Normal(μ, σ) = Normal(; Any[:μ, :σ])
    ((:≪)(::Normal{P, X}, ::Lebesgue{X}) where {P, X}) = true
    ((:≪)(::Lebesgue{X}, ::Normal{P, X}) where {P, X}) = true
end

Note that the eltype function needs to be defined separately by the user.