MeasureTheory

API

MeasureTheory.CorrCholesky

MeasureTheory.LKJCholesky

MeasureBase.logdensity

MeasureTheory.asparams

MeasureTheory.CorrCholesky — Type

CorrCholesky(n)

Cholesky factor of a correlation matrix of size n. Transforms $n(n-1)/2$ real numbers to an $n×n$ lower-triangular matrix L, such that L*L' is a correlation matrix (positive definite, with unit diagonal).

Notes

If

z is a vector of n IID standard normal variates,

σ is an n-element vector of standard deviations,

C is obtained from CorrCholesky(n),

then Diagonal(σ) * C.L * z is a zero-centered multivariate normal variate with the standard deviations σ and correlation matrix C.L * C.U.

MeasureTheory.LKJCholesky — Type

LKJCholesky(k=3, η=1.0)
LKJCholesky(k=3, logη=0.0)

LKJCholesky(k, ...) gives the k×k LKJ distribution (Lewandowski et al 2009) expressed as a Cholesky decomposition. As a special case, for C = rand(LKJCholesky(k=K, η=1.0)) (or equivalently C=rand(LKJCholesky{k}(k=K, logη=0.0))), C.L * C.U is uniform over the set of all K×K correlation matrices. Note, however, that in this case C.L and C.U are not sampled uniformly (because the multiplication is nonlinear).

The logdensity method for this measure applies for LowerTriangular, UpperTriangular, or Diagonal matrices, and will "do the right thing". The logdensity does not check if L*U yields a valid correlation matrix.

Valid values are $η > 0$. When $η > 1$, the distribution is unimodal with a peak at I, while $0 < η < 1$ yields a trough. $η = 2$ is recommended as a vague prior.

Adapted from https://github.com/tpapp/AltDistributions.jl

MeasureBase.logdensity — Function

logdensity(μ::AbstractMeasure [, ν::AbstractMeasure], x::X)

Compute the logdensity of the measure μ at the point x. This is the standard way to define logdensity for a new measure. the base measure is implicit here, and is understood to be basemeasure(μ).

MeasureTheory.asparams — Function

asparams build on TransformVariables.as to construct bijections to the parameter space of a given parameterized measure. Because this is only possible for continuous parameter spaces, we allow constraints to assign values to any subset of the parameters.

asparams(::Type{<:ParameterizedMeasure}, ::Val{::Symbol})

Return a transformation for a particular parameter of a given parameterized measure. For example,

julia> asparams(Normal, Val(:σ))
asℝ₊

asparams(::Type{<: ParameterizedMeasure{N}}, constraints::NamedTuple) where {N}

Return a transformation for a given parameterized measure subject to the named tuple constraints. For example,

julia> asparams(Binomial{(:p,)}, (n=10,))
TransformVariables.TransformTuple{NamedTuple{(:p,), Tuple{TransformVariables.ScaledShiftedLogistic{Float64}}}}((p = as𝕀,), 1)

aspararams(::ParameterizedMeasure)

Return a transformation with no constraints. For example,

julia> asparams(Normal{(:μ,:σ)})
TransformVariables.TransformTuple{NamedTuple{(:μ, :σ), Tuple{TransformVariables.Identity, TransformVariables.ShiftedExp{true, Float64}}}}((μ = asℝ, σ = asℝ₊), 2)